Refactor transport map kernel
